tests: import
Several large part of the project are not covered by these tests. The main
hassle is their interraction with either boost::asio, or their rather
low-level system interraction. The first problem could be solved by
factoring all the interraction to boost::asio in a specific sub-class which
could then be mocked up by the test-suite.

Anyhow, this will need to be re-investigated.